Add Halloween 2025 countdown PHP script
With FPP 9.0+, the OS changed which required a font change. Also, php was updated and a blank default time zone is now treated as UTC instead of local time, so had to add logic to treat DateTime as the local time zone.
---
Introduces a PHP script for FPP to display a countdown timer and show information on matrix overlays for Halloween 2025. The script handles dynamic showtimes, overlay updates, and graceful shutdown on interrupt.